Added code for creating localdate with values
